requirements.txt
kafka-python==1.4.0
ipython
elasticsearch_async
